Add files via upload
Mysite is the name of the project. Inside the Mysite folder there is the "manage.py" file for running the project. Then there are two subfolders namely the "mysite" and "main". The "mysite" subfolder contains the file relevant to the project. The "main" folder contains the files that implements the user registration and log in/out subsystem. A detailed specification of the files will be provided very soon.